feat: 完成em7028的驱动，同时完成测量心率并显示到手表UI上的功能

+++ SoftWare/APP/User/HWInterface/Inc/hw_aht20_interface.h
@@ -9,8 +9,8 @@ typedef struct{
 	device_status_t state;
 	float temperature;
 	float humidity;
-	int (*init)(void);
-	int (*read)(float* temperature,float* humidity);
+	uint8_t (*init)(void);
+	uint8_t (*read)(float* temperature,float* humidity);
 	
 }hw_aht20_interface_t;
